+++ vignettes/grammar-extensions.R
@@ -6,7 +6,7 @@ options(warnPartialMatchArgs = FALSE,
         tibble.print.max = 4,
         tibble.print.min = 4,
         dplyr.summarise.inform = FALSE)
-eval_all <- FALSE  # if TRUE evaluate all code chunks
+eval_all <- TRUE  # if TRUE evaluate all code chunks, which is too slow for checks
 
 ## ---- message=FALSE-----------------------------------------------------------
 library(ggpp)
@@ -17,12 +17,9 @@ eval_ggrepel <- requireNamespace("ggrepel", quietly = TRUE) &&
 if (eval_ggrepel) library(ggrepel)
 eval_magick <- requireNamespace("magick", quietly = TRUE)
 
-## -----------------------------------------------------------------------------
-old_theme <- theme_set(theme_bw())
-
 ## ---- message=FALSE-----------------------------------------------------------
 library(ggpp)
-theme_set(theme_bw() + theme(panel.grid = element_blank()))
+old_theme <- theme_set(theme_bw() + theme(panel.grid = element_blank()))
 
 ## -----------------------------------------------------------------------------
 p1 <- ggplot(mtcars, aes(factor(cyl), mpg)) +
@@ -95,13 +92,13 @@ ggplot(mpg, aes(displ, hwy, colour = factor(cyl))) +
   geom_point() 
 
 ## ---- eval=eval_all-----------------------------------------------------------
-#  tb <- mpg %>%
-#    group_by(cyl) %>%
-#    summarise(hwy = median(hwy), cty = median(cty))
-#  
-#  ggplot(mpg, aes(displ, hwy, colour = factor(cyl))) +
-#    annotate("table", x = 7, y = 44, label = tb) +
-#    geom_point()
+tb <- mpg %>%
+  group_by(cyl) %>%
+  summarise(hwy = median(hwy), cty = median(cty))
+
+ggplot(mpg, aes(displ, hwy, colour = factor(cyl))) +
+  annotate("table", x = 7, y = 44, label = tb) +
+  geom_point() 
 
 ## -----------------------------------------------------------------------------
 ggplot(mpg, aes(displ, hwy, colour = factor(cyl))) +
@@ -170,17 +167,17 @@ ggplot(mpg, aes(displ, hwy, colour = factor(cyl))) +
   theme_bw()
 
 ## ---- eval=eval_all-----------------------------------------------------------
-#  p <- ggplot(mpg, aes(factor(cyl), hwy, fill = factor(cyl))) +
-#    stat_summary(geom = "col", fun = mean, width = 2/3) +
-#    labs(x = "Number of cylinders", y = NULL, title = "Means") +
-#    scale_fill_discrete(guide = "none")
-#  
-#  ggplot(mpg, aes(displ, hwy, colour = factor(cyl))) +
-#    annotate("plot", x = 7, y = 44, label = p + theme_bw(8)) +
-#    geom_point() +
-#    labs(x = "Engine displacement (l)", y = "Fuel use efficiency (MPG)",
-#         colour = "Engine cylinders\n(number)") +
-#    theme_bw()
+p <- ggplot(mpg, aes(factor(cyl), hwy, fill = factor(cyl))) +
+  stat_summary(geom = "col", fun = mean, width = 2/3) +
+  labs(x = "Number of cylinders", y = NULL, title = "Means") +
+  scale_fill_discrete(guide = "none")
+
+ggplot(mpg, aes(displ, hwy, colour = factor(cyl))) +
+  annotate("plot", x = 7, y = 44, label = p + theme_bw(8)) +
+  geom_point() +
+  labs(x = "Engine displacement (l)", y = "Fuel use efficiency (MPG)",
+       colour = "Engine cylinders\n(number)") +
+  theme_bw()
 
 ## ---- eval=eval_magick--------------------------------------------------------
 file.name <- 
@@ -204,10 +201,10 @@ ggplot() +
   theme_bw(12)
 
 ## ---- eval=(eval_all && eval_magick)------------------------------------------
-#  ggplot() +
-#    annotate("grob", x = 1, y = 3, vp.width = 0.5,
-#             label = grid::rasterGrob(image = Isoquercitin, width = 1)) +
-#    theme_bw(12)
+ggplot() +
+  annotate("grob", x = 1, y = 3, vp.width = 0.5,
+           label = grid::rasterGrob(image = Isoquercitin, width = 1)) +
+  theme_bw(12)
 
 ## -----------------------------------------------------------------------------
 ggplot(mpg, aes(displ, hwy, colour = factor(cyl))) +
@@ -224,10 +221,11 @@ my.cars <- my.cars[order(my.cars$wt), ]
 ggplot(my.cars, aes(wt, mpg, label = name)) +
   geom_point() +
   geom_text_s(aes(colour = factor(cyl)),
-                   angle = 90,
-                   hjust = 1.04, nudge_y = -1, 
-                   nudge_x = c(0, 0, -0.05, +0.05, 0, 0 ,0, 0),
-                   arrow = arrow(length = grid::unit(1.5, "mm"))) +
+              vjust = 0.5,
+              angle = 90,
+              nudge_y = -1.5, 
+              nudge_x = c(0, 0, -0.05, +0.05, 0, 0 ,0, 0),
+              arrow = arrow(length = grid::unit(1.5, "mm"))) +
   scale_colour_discrete(l = 40) +
   expand_limits(y = 0)
 
@@ -254,20 +252,20 @@ ggplot(data.tb, aes(x, y)) +
   expand_limits(y = 55, x = 0)
 
 ## ---- eval=(eval_all && eval_magick)------------------------------------------
-#  ggplot(data.tb, aes(x, y)) +
-#    geom_grob_npc(label = list(grid::rasterGrob(image = Robinin, width = 1)),
-#                  npcx = 0.02, npcy = 0.95,
-#                  vp.width = 1/2, vp.height = 1/4) +
-#    geom_point() +
-#    expand_limits(y = 55, x = 0)
+ggplot(data.tb, aes(x, y)) +
+  geom_grob_npc(label = list(grid::rasterGrob(image = Robinin, width = 1)), 
+                npcx = 0.02, npcy = 0.95,
+                vp.width = 1/2, vp.height = 1/4) +
+  geom_point() +
+  expand_limits(y = 55, x = 0)
 
 ## ---- eval=(eval_all && eval_magick)------------------------------------------
-#  ggplot(data.tb, aes(x, y)) +
-#    annotate("grob_npc", label = grid::rasterGrob(image = Robinin, width = 1),
-#                  npcx = 0.02, npcy = 0.95, vp.width = 1/2, vp.height = 1/4) +
-#    geom_point() +
-#    expand_limits(y = 55, x = 0)
-#  
+ggplot(data.tb, aes(x, y)) +
+  annotate("grob_npc", label = grid::rasterGrob(image = Robinin, width = 1), 
+                npcx = 0.02, npcy = 0.95, vp.width = 1/2, vp.height = 1/4) +
+  geom_point() +
+  expand_limits(y = 55, x = 0)
+
 
 ## -----------------------------------------------------------------------------
 corner_letters.tb <- tibble(label = LETTERS[1:4],
@@ -320,22 +318,22 @@ ggplot(my.data, aes(x, y)) +
   expand_limits(y = c(-250, 250))
 
 ## ---- eval=eval_all-----------------------------------------------------------
-#  ggplot(my.data, aes(x, y)) +
-#    geom_quadrant_lines(colour = "red", pool.along = "x") +
-#    stat_quadrant_counts(colour = "red", pool.along = "x") +
-#    geom_point()
+ggplot(my.data, aes(x, y)) +
+  geom_quadrant_lines(colour = "red", pool.along = "x") +
+  stat_quadrant_counts(colour = "red", pool.along = "x") +
+  geom_point()
 
 ## -----------------------------------------------------------------------------
 ggplot(my.data, aes(x, y)) +
   geom_point() +
   stat_quadrant_counts(quadrants = 0L, label.x = "left", 
-                       aes(label = sprintf("%i observations", stat(count))))
+                       aes(label = sprintf("%i observations", after_stat(count))))
 
 ## ---- eval=eval_all-----------------------------------------------------------
-#  ggplot(my.data, aes(x, y)) +
-#    geom_quadrant_lines(colour = "red") +
-#    stat_quadrant_counts(colour = "red", quadrants = c(2, 4)) +
-#    geom_point()
+ggplot(my.data, aes(x, y)) +
+  geom_quadrant_lines(colour = "red") +
+  stat_quadrant_counts(colour = "red", quadrants = c(2, 4)) +
+  geom_point()
 
 ## -----------------------------------------------------------------------------
 ggplot(my.data, aes(x, y, colour = group)) +
@@ -370,21 +368,21 @@ ggplot(data = d, aes(x, y)) +
   stat_dens2d_filter(keep.fraction = 1/4, colour = "red")
 
 ## ---- eval=eval_all-----------------------------------------------------------
-#  ggplot(data = d, aes(x, y)) +
-#    geom_point() +
-#    stat_dens2d_filter(keep.fraction = 1/4, keep.number = 50, colour = "red")
+ggplot(data = d, aes(x, y)) +
+  geom_point() +
+  stat_dens2d_filter(keep.fraction = 1/4, keep.number = 50, colour = "red")
 
 ## ---- eval=eval_all-----------------------------------------------------------
-#  ggplot(data = d, aes(x, y)) +
-#    geom_point() +
-#    stat_dens2d_filter(keep.fraction = 1, keep.number = 50, colour = "red")
+ggplot(data = d, aes(x, y)) +
+  geom_point() +
+  stat_dens2d_filter(keep.fraction = 1, keep.number = 50, colour = "red")
 
 ## ---- eval=eval_all-----------------------------------------------------------
-#  ggplot(data = d, aes(x, y, colour = group)) +
-#     stat_dens2d_filter(keep.fraction = 0.25,
-#                        size = 3,
-#                        colour = "black") +
-#     geom_point()
+ggplot(data = d, aes(x, y, colour = group)) +
+   stat_dens2d_filter(keep.fraction = 0.25,
+                      size = 3,
+                      colour = "black") +
+   geom_point()
 
 ## -----------------------------------------------------------------------------
 ggplot(data = d, aes(x + rep(c(-2,2), rep(50,2)), 
@@ -401,23 +399,23 @@ ggplot(data = d, aes(x + rep(c(-2,2), rep(50,2)),
                       keep.fraction = 0.25)
 
 ## ---- eval=eval_all-----------------------------------------------------------
-#  ggplot(data = d, aes(x, y, label = lab, colour = group)) +
-#    stat_dens2d_labels(keep.fraction = 1/10,
-#                       hjust = "outward", vjust = "outward") +
-#    geom_point()
+ggplot(data = d, aes(x, y, label = lab, colour = group)) +
+  stat_dens2d_labels(keep.fraction = 1/10, 
+                     hjust = "outward", vjust = "outward") +
+  geom_point()
 
 ## ---- eval=eval_ggrepel-------------------------------------------------------
-#  ggplot(data = d, aes(x, y, label = lab, colour = group)) +
-#    geom_point() +
-#    stat_dens2d_labels(geom = "text_repel",
-#                       keep.fraction = 0.45)
+ggplot(data = d, aes(x, y, label = lab, colour = group)) +
+  geom_point() +
+  stat_dens2d_labels(geom = "text_repel", 
+                     keep.fraction = 0.45)
 
 ## ---- eval=(eval_all && eval_ggrepel)-----------------------------------------
-#  ggplot(data = d, aes(x, y, label = lab, colour = group)) +
-#    stat_dens2d_labels(geom = "label_repel",
-#                       keep.fraction = 0.2,
-#                       label.fill = NA) +
-#      geom_point()
+ggplot(data = d, aes(x, y, label = lab, colour = group)) +
+  stat_dens2d_labels(geom = "label_repel", 
+                     keep.fraction = 0.2, 
+                     label.fill = NA) +
+    geom_point()
 
 ## ---- eval = FALSE------------------------------------------------------------
 #  random_string <- function(len = 6) {
@@ -440,21 +438,21 @@ ggplot(data = d, aes(x, y)) +
                      colour = "red")
 
 ## ---- eval=eval_all-----------------------------------------------------------
-#  ggplot(data = d, aes(x, y)) +
-#    geom_point() +
-#    stat_dens1d_filter(keep.fraction = 0.25,
-#                       colour = "red",
-#                       orientation = "y")
+ggplot(data = d, aes(x, y)) +
+  geom_point() +
+  stat_dens1d_filter(keep.fraction = 0.25,
+                     colour = "red",
+                     orientation = "y")
 
 ## ---- eval=eval_ggrepel-------------------------------------------------------
-#  ggplot(data = d, aes(x, y, label = lab, colour = group)) +
-#    geom_point() +
-#    stat_dens1d_labels(geom = "text_repel")
+ggplot(data = d, aes(x, y, label = lab, colour = group)) +
+  geom_point() +
+  stat_dens1d_labels(geom = "text_repel")
 
 ## ---- eval=(eval_all && eval_ggrepel)-----------------------------------------
-#  ggplot(data = d, aes(x, y, label = lab, colour = group)) +
-#    geom_point() +
-#    stat_dens1d_labels(geom = "text_repel", orientation = "y")
+ggplot(data = d, aes(x, y, label = lab, colour = group)) +
+  geom_point() +
+  stat_dens1d_labels(geom = "text_repel", orientation = "y")
 
 ## -----------------------------------------------------------------------------
 set.seed(84532)
@@ -468,29 +466,27 @@ df <- data.frame(
 ## -----------------------------------------------------------------------------
 ggplot(df, aes(x, y, label = l)) +
   geom_point() +
-  geom_text_s(position = position_nudge_keep(x = 0.1),
-              hjust = "left") +
+  geom_text_s(position = position_nudge_keep(x = 0.1)) +
   expand_limits(x = 2.5)
 
 ## -----------------------------------------------------------------------------
 ggplot(df, aes(x, y, label = l)) +
   geom_point() +
-  geom_text_s(position = position_nudge_keep(x = 0.1),
-              hjust = "left", add.segments = FALSE) +
+  geom_text_s(position = position_nudge_keep(x = 0.1), add.segments = FALSE) +
   expand_limits(x = 2.5)
 
 ## ---- eval=eval_ggrepel-------------------------------------------------------
-#  ggplot(df, aes(x, y, label = l)) +
-#    geom_point() +
-#    geom_text_repel(position = position_nudge(x = 0.3),
-#                    min.segment.length = 0,
-#                    max.iter = 0)
+ggplot(df, aes(x, y, label = l)) +
+  geom_point() +
+  geom_text_repel(position = position_nudge(x = 0.3),
+                  min.segment.length = 0, 
+                  max.iter = 0)
 
 ## ---- eval=eval_ggrepel-------------------------------------------------------
-#  ggplot(df, aes(x, y, label = l)) +
-#    geom_point() +
-#    geom_text_repel(position = position_nudge_keep(x = 0.3),
-#                    min.segment.length = 0, max.iter = 0)
+ggplot(df, aes(x, y, label = l)) +
+  geom_point() +
+  geom_text_repel(position = position_nudge_keep(x = 0.3),
+                  min.segment.length = 0, max.iter = 0)
 
 ## -----------------------------------------------------------------------------
 ggplot(df, aes(x, y, label = l)) +
@@ -503,117 +499,113 @@ ggplot(df, aes(x, y, label = ifelse(x < 1, "", l) )) +
   geom_text_s(position = position_nudge_to(y = 2.3),
               segment.color = "red",
               arrow = arrow(length = unit(0.015, "npc")),
-              angle = 90,
-              hjust = "left") +
+              angle = 90) +
   expand_limits(x = 3)
 
 ## ---- eval=eval_ggrepel-------------------------------------------------------
-#  ggplot(df, aes(x, y, label = l)) +
-#    geom_point(color = "red", size = 3) +
-#    geom_text_repel(position = position_nudge_to(y = -2.5),
-#                    size = 3,
-#                    color = "red",
-#                    angle = 90,
-#                    hjust = 0,
-#                    box.padding = 0.1,
-#                    segment.color = NA,
-#                    direction = "x") +
-#    geom_rug(sides = "b", length = unit(0.02, "npc"), color = "red")
+ggplot(df, aes(x, y, label = l)) +
+  geom_point(color = "red", size = 3) +
+  geom_text_repel(position = position_nudge_to(y = -2.5), 
+                  size = 3,
+                  color = "red",
+                  angle = 90,
+                  hjust = 0,
+                  box.padding = 0.1,
+                  segment.color = NA,
+                  direction = "x") +
+  geom_rug(sides = "b", length = unit(0.02, "npc"), color = "red")
+
+## -----------------------------------------------------------------------------
+ggplot(df, aes(x, y, label = l)) +
+  geom_point() +
+  geom_text_s(position = 
+                    position_nudge_center(x = -0.1, center_x = 0))
+
+## -----------------------------------------------------------------------------
+ggplot(df, aes(x, y, label = l)) +
+  geom_point() +
+  geom_text_s(position = position_nudge_center(x = 0.1, direction = "split")) +
+  expand_limits(x = c(-3, 3))
+
+## -----------------------------------------------------------------------------
+ggplot(df, aes(x, y, label = l)) +
+  geom_point() +
+  geom_text_s(position = 
+                position_nudge_center(x = 0.1, 
+                                      center_x = 1, 
+                                      direction = "split")) +
+  expand_limits(x = c(-3, 3))
+
+## -----------------------------------------------------------------------------
+ggplot(df, aes(x, y, label = l)) +
+  geom_point() +
+  geom_text_s(position = position_nudge_center(x = 0.1, 
+                                               center_x = median, 
+                                               direction = "split")) +
+  expand_limits(x = c(-3, 3))
+
+## -----------------------------------------------------------------------------
+ggplot(df, aes(x, y, label = l)) +
+  geom_point() +
+  geom_text_s(position = 
+                    position_nudge_center(x = 0.1,
+                                          center_x = function(x) {
+                                            quantile(x, 
+                                                     probs = 1/4, 
+                                                     names = FALSE)
+                                          },
+                                          direction = "split")) +
+  expand_limits(x = c(-3, 3))
 
 ## ---- eval=eval_ggrepel-------------------------------------------------------
-#  ggplot(df, aes(x, y, label = l)) +
-#    geom_point() +
-#    geom_text_repel(position =
-#                      position_nudge_center(x = 0.3, center_x = 0),
-#                      min.segment.length = 0, max.iter = 0)
+ggplot(df, aes(x, y, label = l)) +
+  geom_point() +
+  geom_text_repel(angle = 90,
+                  position = position_nudge_center(y = 1,
+                                                   direction = "split")) +
+  expand_limits(y = c(-2, 7))
 
 ## ---- eval=eval_ggrepel-------------------------------------------------------
-#  ggplot(df, aes(x, y, label = l)) +
-#    geom_point() +
-#    geom_text_repel(position =
-#                      position_nudge_center(x = 0.3,
-#                                            direction = "split"),
-#                    min.segment.length = 0)
+ggplot(df, aes(x, y, label = l)) +
+  geom_point() +
+  geom_text_repel(position = position_nudge_center(x = 0.3,
+                                                   y = 0.3,
+                                                   center_x = -0.5,
+                                                   direction = "split"),
+                  min.segment.length = 0) +
+  expand_limits(x = c(-3, 3))
 
 ## ---- eval=eval_ggrepel-------------------------------------------------------
-#  ggplot(df, aes(x, y, label = l)) +
-#    geom_point() +
-#    geom_text_repel(position =
-#                      position_nudge_center(x = 0.3,
-#                                            center_x = 1,
-#                                            direction = "split"),
-#                    min.segment.length = 0)
+ggplot(df, aes(x, y, label = l)) +
+  geom_point() +
+  geom_text_repel(position = 
+                    position_nudge_center(x = 0.25,
+                                          y = 0.4,
+                                          direction = "radial"),
+                  min.segment.length = 0)
 
 ## ---- eval=eval_ggrepel-------------------------------------------------------
-#  ggplot(df, aes(x, y, label = l)) +
-#    geom_point() +
-#    geom_text_repel(position =
-#                      position_nudge_center(x = 0.3,
-#                                            center_x = median,
-#                                            direction = "split"),
-#                    min.segment.length = 0)
+ggplot(df, aes(x, y, label = l)) +
+  geom_point() +
+  geom_text_repel(position = position_nudge_center(x = 0.4,
+                                                   y = 0.4,
+                                                   center_x = -0.5,
+                                                   center_y = 1,
+                                                   direction = "radial"),
+                  min.segment.length = 0) +
+  expand_limits(x = c(-3, 3))
 
 ## ---- eval=eval_ggrepel-------------------------------------------------------
-#  ggplot(df, aes(x, y, label = l)) +
-#    geom_point() +
-#    geom_text_repel(position =
-#                      position_nudge_center(x = 0.3,
-#                                            center_x = function(x) {
-#                                              quantile(x,
-#                                                       probs = 1/4,
-#                                                       names = FALSE)
-#                                            },
-#                                            direction = "split"),
-#                    min.segment.length = 0)
-
-## ---- eval=eval_ggrepel-------------------------------------------------------
-#  ggplot(df, aes(x, y, label = l)) +
-#    geom_point() +
-#    geom_text_repel(angle = 90,
-#                    position =
-#                      position_nudge_center(y = 0.1,
-#                                            direction = "split"))
-
-## ---- eval=eval_ggrepel-------------------------------------------------------
-#  ggplot(df, aes(x, y, label = l)) +
-#    geom_point() +
-#    geom_text_repel(position =
-#                      position_nudge_center(x = 0.1,
-#                                            y = 0.15,
-#                                            direction = "split"))
-
-## ---- eval=eval_ggrepel-------------------------------------------------------
-#  ggplot(df, aes(x, y, label = l)) +
-#    geom_point() +
-#    geom_text_repel(position =
-#                      position_nudge_center(x = 0.25,
-#                                            y = 0.4,
-#                                            direction = "radial"),
-#                    min.segment.length = 0)
-
-## ---- eval=eval_ggrepel-------------------------------------------------------
-#  ggplot(df, aes(x, y, label = l)) +
-#    geom_point() +
-#    geom_text_repel(position =
-#                      position_nudge_center(x = 0.15,
-#                                            y = 0.3,
-#                                            center_x = 0,
-#                                            center_y = 0,
-#                                            direction = "radial"),
-#                    min.segment.length = 0,
-#                    hjust = "outward", vjust = "outward")
-
-## ---- eval=eval_ggrepel-------------------------------------------------------
-#  ggplot(df, aes(x, y, label = l)) +
-#    geom_point() +
-#    geom_text_repel(position =
-#                      position_nudge_center(x = 0.2,
-#                                            center_x = 0,
-#                                            direction = "split"),
-#                    aes(hjust = ifelse(x < 0, 1, 0)),
-#                    direction = "y",
-#                    min.segment.length = 0) +
-#    expand_limits(x = c(-3, 3))
+ggplot(df, aes(x, y, label = l)) +
+  geom_point() +
+  geom_text_repel(position = 
+                    position_nudge_center(x = 0.2,
+                                          center_x = 0,
+                                          direction = "split"),
+                  aes(hjust = ifelse(x < 0, 1, 0)),
+                  direction = "y",
+                  min.segment.length = 0) +
+  expand_limits(x = c(-3, 3))
 
 ## -----------------------------------------------------------------------------
 set.seed(16532)
@@ -634,7 +626,7 @@ ggplot(df, aes(x, 2 * x, label = l)) +
 ## -----------------------------------------------------------------------------
 ggplot(subset(df, x >= 0), aes(x, yyy)) +
   geom_point() +
-  stat_smooth(method = "lm") +
+  stat_smooth(method = "lm", formula = y ~ x) +
   geom_text(aes(label = l),
             vjust = "center", hjust = "center",
             position = position_nudge_line(x = 0, y = 1.2,
@@ -673,50 +665,51 @@ ggplot(df, aes(x, y, label = l)) +
   geom_text(position = position_nudge_line(x = -0.6, y = -6))
 
 ## -----------------------------------------------------------------------------
-ggplot(df, aes(x, yy, label = l)) +
+ggplot(df, aes(x, yy)) +
   geom_point() +
-  stat_smooth() +
-  geom_text(aes(y = yy, label = l),
+  stat_smooth(method = "loess", formula = y ~ x) +
+  geom_text(aes(label = l),
             position = position_nudge_line(x = 0.6, 
                                            y = 6,
                                            direction = "split"))
 
 ## ---- eval=eval_ggrepel-------------------------------------------------------
-#  ggplot(df, aes(x, yy, label = l)) +
-#    geom_point() +
-#    stat_smooth() +
-#    geom_label_repel(aes(y = yy, label = paste("point", l)),
-#              position = position_nudge_line(x = 0.6,
-#                                             y = 6,
-#                                             direction = "split"),
-#              min.segment.length = 0)
+ggplot(df, aes(x, yy)) +
+  geom_point() +
+  stat_smooth(method = "loess", formula = y ~ x) +
+  geom_label_s(aes(label = l),
+                   position = position_nudge_line(x = 0.4, 
+                                                  y = 4,
+                                                  direction = "split")) +
+  expand_limits(y = -12)
 
 ## -----------------------------------------------------------------------------
-ggplot(df, aes(x, yy, label = l)) +
+ggplot(df, aes(x, yy)) +
   geom_point() +
   stat_smooth(method = "lm", formula = y ~ poly(x, 2, raw = TRUE)) +
-  geom_text(aes(y = yy, label = l),
-            position = position_nudge_line(method = "lm",
-                                           x = 0.5, 
-                                           y = 5,
-                                           formula = y ~ poly(x, 2, raw = TRUE),
-                                           direction = "split"))
+  geom_text_s(aes(label = l),
+              position = position_nudge_line(method = "lm",
+                                             x = 0.5, 
+                                             y = 5,
+                                             formula = y ~ poly(x, 2, raw = TRUE),
+                                             direction = "split"))
 
 ## ---- eval=eval_ggrepel-------------------------------------------------------
-#  ggplot(df, aes(x, yy, label = l)) +
-#    geom_point() +
-#    stat_smooth(method = "lm", formula = y ~ poly(x, 2, raw = TRUE)) +
-#    geom_text_repel(aes(y = yy, label = l),
-#                    box.padding = 0.5,
-#                    min.segment.length = Inf)
+ggplot(df, aes(x, yy)) +
+  geom_point() +
+  stat_smooth(method = "lm", formula = y ~ poly(x, 2, raw = TRUE)) +
+  geom_text_repel(aes(label = l),
+                  box.padding = 0.5,
+                  min.segment.length = Inf)
 
 ## ---- eval=eval_ggrepel-------------------------------------------------------
-#  ggplot(df, aes(x, yy, label = l)) +
-#    geom_point() +
-#    stat_smooth() +
-#    geom_label_repel(aes(y = yy, label = paste("point", l)),
-#                    box.padding = 0.5,
-#                    min.segment.length = 0)
+ggplot(df, aes(x, yy)) +
+  geom_point() +
+  stat_smooth(method = "loess", formula = y ~ x) +
+  geom_label_repel(aes(label = paste("point", l)),
+                  box.padding = 0.5,
+                  min.segment.length = 0) +
+  expand_limits(y = -10, x = c(-13, 13))
 
 ## -----------------------------------------------------------------------------
 df <- data.frame(x = rep(1:10, 2),
@@ -754,6 +747,87 @@ ggplot(df, aes(x, y, label = l)) +
   coord_equal(ratio = 1)
 
 ## -----------------------------------------------------------------------------
+ df <- data.frame(x1 = c(1, 2, 1, 3, -1),
+                  x2 = c("a", "a", "b", "b", "b"),
+                  grp = c("A", "B", "C", "D", "E"))
+
+## -----------------------------------------------------------------------------
+ggplot(data = df, aes(x1, x2, group = grp)) +
+   geom_col(aes(fill = grp), width = 0.8,
+            position = position_dodge()) +
+   geom_vline(xintercept = 0) +
+   geom_text(
+     aes(label = grp),
+     position = position_dodgenudge(x = 0.09, direction = "split", width = 0.8)) +
+   theme(legend.position = "none")
+
+## -----------------------------------------------------------------------------
+ ggplot(data = df, aes(x2, x1, group = grp)) +
+   geom_col(aes(fill = grp), width = 0.75,
+            position = position_dodge(width = 0.75)) +
+   geom_hline(yintercept = 0) +
+   geom_text(aes(label = grp),
+             position = position_dodgenudge(y = 0.1,
+                                            direction = "split",
+                                            width = 0.75),
+             size = 3) +
+   theme(legend.position = "none")
+
+## -----------------------------------------------------------------------------
+ ggplot(data = df, aes(x1, x2, group = grp)) +
+   geom_col(aes(fill = grp), width = 0.5) +
+   geom_vline(xintercept = 0) +
+   geom_text(
+     aes(label = grp),
+     position = position_stacknudge(vjust = 0.5, y = 0.33)) +
+   theme(legend.position = "none")
+
+## -----------------------------------------------------------------------------
+ ggplot(data = subset(df, x1 >= 0), aes(x2, x1, group = grp)) +
+   geom_col(aes(fill = grp), width=0.5, position = position_fill()) +
+   geom_vline(xintercept = 0) +
+   geom_text(
+     aes(label = grp),
+     position = position_fillnudge(vjust = 1, y = -0.05)) +
+   theme(legend.position = "none")
+
+
+## -----------------------------------------------------------------------------
+ jitter <- position_jitter(width = 0.2, height = 2, seed = 123)
+
+ jitter_nudge <- position_jitternudge(width = 0.2, height = 2,
+                                      seed = 123, x = 0.1,
+                                      direction = "split",
+                                      nudge.from = "jittered")
+ ggplot(mpg[1:20, ],
+        aes(cyl, hwy, label = drv)) +
+   geom_point(position = jitter) +
+   geom_text_s(position = jitter_nudge)
+
+
+## -----------------------------------------------------------------------------
+jitter <- position_jitter(width = 0.2, height = 2, seed = 123)
+
+jitter_nudge <- position_jitternudge(width = 0.2, height = 2,
+                                      seed = 123, x = 0.35,
+                                      direction = "split",
+                                      nudge.from = "original.x")
+ ggplot(mpg[1:20, ],
+        aes(cyl, hwy, label = drv)) +
+   geom_point(position = jitter) +
+   geom_text_s(position = jitter_nudge)
+
+## -----------------------------------------------------------------------------
+ ggplot(mpg[1:20, ],
+        aes(cyl, hwy, label = drv)) +
+   geom_point() +
+   geom_point_s(position =
+                position_jitter_keep(width = 0.3, height = 2, seed = 123),
+                color = "red",
+                arrow = grid::arrow(length = unit(0.4, "lines")))
+
+
+## -----------------------------------------------------------------------------
 make_data_tbl <- function(nrow = 100, rfun = rnorm, ...) {
   if (nrow %% 2) {
     nrow <- nrow + 1
@@ -769,31 +843,31 @@ make_data_tbl <- function(nrow = 100, rfun = rnorm, ...) {
 }
 
 ## ---- eval=eval_all-----------------------------------------------------------
-#  ggplot(data = make_data_tbl(300), aes(x, y)) +
-#    geom_point() +
-#    stat_dens2d_filter(colour = "red",
-#                       keep.sparse = FALSE,
-#                       keep.fraction = 1/3)
+ggplot(data = make_data_tbl(300), aes(x, y)) +
+  geom_point() +
+  stat_dens2d_filter(colour = "red", 
+                     keep.sparse = FALSE, 
+                     keep.fraction = 1/3)
 
 ## ---- eval=eval_all-----------------------------------------------------------
-#  ggplot(data = make_data_tbl(300), aes(x, y)) +
-#    geom_point() +
-#    stat_dens2d_filter(colour = "red",
-#                       keep.sparse = FALSE,
-#                       keep.fraction = 1/3)+
-#    stat_dens2d_filter(colour = "blue",
-#                       keep.fraction = 1/3)
+ggplot(data = make_data_tbl(300), aes(x, y)) +
+  geom_point() +
+  stat_dens2d_filter(colour = "red", 
+                     keep.sparse = FALSE, 
+                     keep.fraction = 1/3)+
+  stat_dens2d_filter(colour = "blue", 
+                     keep.fraction = 1/3)
 
 ## ---- eval=eval_all-----------------------------------------------------------
-#  ggplot(data = make_data_tbl(300, rfun = runif), aes(x, y)) +
-#    geom_point() +
-#    stat_dens2d_filter(colour = "red", keep.fraction = 1/2)
+ggplot(data = make_data_tbl(300, rfun = runif), aes(x, y)) +
+  geom_point() +
+  stat_dens2d_filter(colour = "red", keep.fraction = 1/2)
 
 ## ---- eval=eval_all-----------------------------------------------------------
-#  ggplot(data = make_data_tbl(300, rfun = rgamma, shape = 2),
-#         aes(x, y)) +
-#    geom_point() +
-#    stat_dens2d_filter(colour = "red", keep.fraction = 1/3)
+ggplot(data = make_data_tbl(300, rfun = rgamma, shape = 2), 
+       aes(x, y)) +
+  geom_point() +
+  stat_dens2d_filter(colour = "red", keep.fraction = 1/3)
 
 ## -----------------------------------------------------------------------------
 class(austres)
@@ -806,35 +880,35 @@ austres.df <- try_tibble(austres, as.numeric = TRUE)
 head(austres.df, 4)
 
 ## ---- eval=eval_all-----------------------------------------------------------
-#  class(lynx)
-#  lynx.df <- try_tibble(lynx)
-#  class(lynx.df)
-#  head(lynx.df, 3)
+class(lynx)
+lynx.df <- try_tibble(lynx)
+class(lynx.df)
+head(lynx.df, 3)
 
 ## -----------------------------------------------------------------------------
 lynx.df <- try_tibble(lynx, "year")
 head(lynx.df, 3)
 
 ## ---- eval=eval_all-----------------------------------------------------------
-#  lynx_n.df <- try_tibble(lynx, "year", as.numeric = TRUE)
-#  lapply(lynx_n.df, "class")
-#  head(lynx_n.df, 3)
+lynx_n.df <- try_tibble(lynx, "year", as.numeric = TRUE)
+lapply(lynx_n.df, "class")
+head(lynx_n.df, 3)
 
 ## ---- eval=eval_all-----------------------------------------------------------
-#  try_tibble(1:5)
+try_tibble(1:5)
 
 ## ---- eval=eval_all-----------------------------------------------------------
-#  try_tibble(letters[1:5])
+try_tibble(letters[1:5])
 
 ## ---- eval=eval_all-----------------------------------------------------------
-#  try_tibble(factor(letters[1:5]))
+try_tibble(factor(letters[1:5]))
 
 ## ---- eval=eval_all-----------------------------------------------------------
-#  try_tibble(list(x = rep(1,5), y = 1:5))
+try_tibble(list(x = rep(1,5), y = 1:5))
 
 ## ---- eval=eval_all-----------------------------------------------------------
-#  try_tibble(data.frame(x = rep(1,5), y = 1:5))
+try_tibble(data.frame(x = rep(1,5), y = 1:5))
 
 ## ---- eval=eval_all-----------------------------------------------------------
-#  try_tibble(matrix(1:10, ncol = 2))
+try_tibble(matrix(1:10, ncol = 2))